Protect specific routes for admin

+++ src/app/Http/Controllers/CarModelController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\CarModel;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\View\View;
 
 class CarModelController extends Controller
@@ -24,12 +25,18 @@ class CarModelController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create(): View
+    public function create(): View|RedirectResponse
     {
-        $viewData = [];
-        $viewData['title'] = 'Car Models - EasyCar';
+        $user = Auth::user();
 
-        return view('carModel.create')->with('viewData', $viewData);
+        if (! $user->isAdmin()) {
+            return redirect('/unauthorized');
+        } else {
+            $viewData = [];
+            $viewData['title'] = 'Car Models - EasyCar';
+
+            return view('carModel.create')->with('viewData', $viewData);
+        }
     }
 
     /**
@@ -37,10 +44,16 @@ class CarModelController extends Controller
      */
     public function save(Request $request): RedirectResponse
     {
-        CarModel::validate($request);
-        CarModel::create($request->only(['brand', 'model', 'description']));
+        $user = Auth::user();
 
-        return back()->with('status', 'successfully created');
+        if (! $user->isAdmin()) {
+            return redirect('/unauthorized');
+        } else {
+            CarModel::validate($request);
+            CarModel::create($request->only(['brand', 'model', 'description']));
+
+            return back()->with('status', 'successfully created');
+        }
     }
 
     /**
@@ -62,8 +75,14 @@ class CarModelController extends Controller
      */
     public function delete(string $id): RedirectResponse
     {
-        CarModel::destroy($id);
+        $user = Auth::user();
 
-        return redirect('car-model');
+        if (! $user->isAdmin()) {
+            return redirect('/unauthorized');
+        } else {
+            CarModel::destroy($id);
+
+            return redirect('car-model');
+        }
     }
 }
